domain.MessageReply->application.BotMessage.
The class has nothing to do with the domain: at the impersonator's
level, there is no interface with the user, no reply to messages etc
(the bot is an application level concern).

Also, MessageReply is a little vague, it could mean any kind of reply to
a message (with the "reply" feature most chat apps have). BotMessage
encompasses way better what this class is about: a message from the bot
to the user.