Multi language support (#2)
* Add tests to Config class.

* Remove unused private method in Config class.

* Make env vars testable with a FunctionalInterface in Config.

* Complete Config testing.

* Add language in Config. Just FR for now, by default.

* domain.MessageReply->application.BotMessage.

The class has nothing to do with the domain: at the impersonator's
level, there is no interface with the user, no reply to messages etc
(the bot is an application level concern).

Also, MessageReply is a little vague, it could mean any kind of reply to
a message (with the "reply" feature most chat apps have). BotMessage
encompasses way better what this class is about: a message from the bot
to the user.

* ApplicationMessageProvider, with a first InMemory implementation.

* Remove useless MeCommand, use LikeCommand.

* Handle UserNotFound error case directly in LikeCommand.

All in all, it seems more relevant there, and also this is setting the
ground for the ApplicationMessageProvider centralised usage from
commands.

* CocoChatBotApplication now replies to the message through an IncomingMessage interface.

This puts less strain on Messaging App (Discord, Telegram, etc) specific
implementation, by having the generic CocoChatBotApplication directly
replying to the message.

* Refactor commands.

Commands are now more generic: any operation on the domain is a command,
including adding messages to it (RegisterMessageCommand).

apply() now takes 0 parameters; commands that need the impersonator
receive it from the parser at creation time.

* Use ApplicationMessageProvider in commands when needed, to get localized message for frontend.

* Read LANGUAGE env var and use it to set the bot language.

* Support english language, and make it default language.

Support both english and french. For now, the in memory InMemoryApplicationMessageProvider
is enough, we have very few messages, a proper yaml config files
implementation would be overkill.

* Add Configuration section to README

* Rename Discord classes for more clarity.

DiscordChatBotService -> DiscordMessageListener: more precise and accurate.
DirectAccessMessagesSource -> DiscordDirectAccessMessagesSource: more precise.

* Tick boxes in README