Load messages asynchronously, to take advantage of Flux multi-threading.
In order to add concurrently to the Impersonator, rework it to use
threadsafe ConcurrentMap, and AtomicInteger.

Then, the addAllMessagesFromSource method uses a scheduler to make
subscribers work on multiple threads. Add a onComplete runnable to tell
when job is done. Return the Flux instead of subscribing to it, so it
can be run differently in main and test contexts.

Will probably have to rethink the addAllMessagesFromSource logic to get
it out of the Impersonator interface, those default methods are a lil
ugly... Might have to introduce a Bot in domain... or extract the addAll
logic to the application layer, leave only addMessage() in domain?